aaaaaaaaaaa

+++ include/ble.h
@@ -4,14 +4,13 @@
 #include <BLEServer.h>
 #include <BLEUtils.h>
 
-// Предварительное объявление классов колбэков,
-// чтобы мы могли объявить их "друзьями" (friend) класса Ble
+
 class ServerCallbacks;
 class RXCallback;
 
 class Ble {
 private:
-    // UUID лучше держать константами здесь или в cpp
+
     const char* SERVICE_UUID = "6e400001-b5a3-f393-e0a9-e50e24dcca9e";
     const char* CHARACTERISTIC_RX = "6e400002-b5a3-f393-e0a9-e50e24dcca9e";
 
@@ -23,14 +22,13 @@ private:
 
     BLEServer* pServer = nullptr;
 
-    // Объявляем эти классы друзьями, чтобы они могли менять приватные переменные (targetLat и т.д.)
+
     friend class ServerCallbacks;
     friend class RXCallback;
 
 public:
-    void waitForTargetCoords();
-    
-    // Эти функции должны возвращать float, а не void
-    float getTargetLat();
-    float getTargetLon();
+    void waitForTargetCoords(); // блокирует выполнение, пока не будут получены координаты
+
+    double getTargetLat(); // возвращает полученную широту
+    double getTargetLon(); // возвращает полученную долготу
 };
